Removed remaining moved files and fixed the benchmarks project issue
