+++ benchmarks/ProtoActorBenchmarks/LongBenchmark.cs
@@ -2,9 +2,10 @@ using System;
 using System.Linq;
 using System.Threading.Tasks;
 using BenchmarkDotNet.Attributes;
+using Proto;
 using Proto.Mailbox;
 
-namespace Proto.Actor.Benchmarks
+namespace ProtoActorBenchmarks
 {
     [MemoryDiagnoser, InProcess]
     public class LongBenchmark
