+++ benchmarks/ProtoActorBenchmarks/ProtoActorBenchmarks.csproj
@@ -2,16 +2,11 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>netcoreapp3.1</TargetFramework>
+    <ServerGarbageCollection>true</ServerGarbageCollection>
+    <ConcurrentGarbageCollection>true</ConcurrentGarbageCollection>
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="BenchmarkDotNet" Version="0.12.0" />
-  </ItemGroup>
-  <ItemGroup>
-    <PackageReference Remove="AutoFixture.Xunit2"/>
-    <PackageReference Remove="FluentAssertions" />
-    <PackageReference Remove="Microsoft.NET.Test.Sdk"/>
-    <PackageReference Remove="xunit" />
-    <PackageReference Remove="xunit.runner.visualstudio" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\src\Proto.Actor\Proto.Actor.csproj" />
