--- src/ClusterProviders/Proto.Cluster.SingleRemoteInstance/Proto.Cluster.SingleRemoteInstance.csproj
@@ -1,10 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk" ToolsVersion="15.0">
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <IsPackable>true</IsPackable>
-  </PropertyGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\..\Proto.Actor\Proto.Actor.csproj"/>
-    <ProjectReference Include="..\..\Proto.Cluster\Proto.Cluster.csproj"/>
-  </ItemGroup>
-</Project>
