--- src/ClusterProviders/Proto.Cluster.SingleRemoteInstance/Protos.proto
@@ -1,8 +1,0 @@
-﻿syntax = "proto3";
-package singleremoteinstance;
-option csharp_namespace = "Proto.Cluster.SingleRemoteInstance";
-
-message GetKinds {}
-message GetKindsResponse {
-    repeated string Kinds=1;
-}
