--- src/ClusterProviders/Proto.Cluster.SingleRemoteInstance/build.bat
@@ -1,1 +1,0 @@
-%userprofile%\.nuget\packages\Grpc.Tools\1.10.0\tools\windows_x64\protoc Protos.proto -I=. -I=../../../src --csharp_out=. --csharp_opt=file_extension=.g.cs --grpc_out . --plugin=protoc-gen-grpc=%userprofile%\.nuget\packages\Grpc.Tools\1.10.0\tools\windows_x64\grpc_csharp_plugin.exe
